+++ vue-front/src/views/About.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <p></p>
+    <img
+      src="../../public/jiro.jpg"
+      alt="ラーメン296"
+      title="ラーメン296"
+      width="30%"
+      height="30%"
+      class="logo offset-lg-4"
+    />
+    <h1 class="offset-lg-4">ラーメン二郎 is God!!</h1>
+  </div>
+</template>
+
